+ add endpoints for entity editing ~ fix CommonText
New GET and POST endpoints /entities/{id}/edit for authenticated users
has been added. Also a new EntityEdit component, which contains
all translations for the user-editable field requested on thos
endpoints.

Also the CommonText component has been remodeled into a CommonTextList
component and a CommonText object component. The former is used
for the fields in Entity, Activity and Album as a list of CommonText
objects. This provides more flexibility regarding examples of lists,
while still retaining the initial idea that the CommonText is a single
object either containing a CommonTextString, a CommonTextArray or
a CommonTextObject in its `data` property.